Update the checking procedure for teradata migration (#20)
Added / updated the checking procedure for the Teradata migration guide.
When the script parameter CHECK_MIGRATION is set to true, checking tables for each source table will be created in the migrated schema with '_MIG_CHK' as a suffix. Those tables will be filled with the results of standardized test queries executed by Teradata and Exasol.
To generate an overview of all the checking tables, a table with the name of the source schema and the suffix '_MIG_CHK' will be created in the database migration schema, to allow for easy analysis.