Update oracle_to_exasol.sql
Updte of the oracle to exasol database migration script.
The update contains script parameters that control the generation of primary and foreign keys, degree of parallelism for the data migration, checking tables for migration testing and some other improvements.

The option for parallel statements defines the degree of parallelism for the import statements. If a table is partitioned, the number of rows per partition will be determined to distribute the partions across the parallel statements evenly. If the table does not have partitions, the ora_hash function on the row_id will be used to assign the rows to one of the parallel statements.

The script options for primary / foreign keys control the code being generated. The statements will be generated anyway. When the option is not set (FALSE) the statements will be commented out.

The script also offers the option to create checking tables. These tables are generated for every table in the scope. They will be created in the migrated schema. The name of the checking tables are the same as the orginal ones with the suffix "_MIG_CHK".
A summary table is created in the database_migration schema of the script with the name of the schema and the suffix "_MIG_CHK", allowing to query all metric for all tables of the scope easily to confirm a complete, correct and successful migration.